+++ CHANGELOG.rst
@@ -17,6 +17,18 @@ Unreleased
 ----------
 * nothing unreleased
 
+[6.6.8] - 2026-03-05
+---------------------
+* feat: Add enterprise admin invite via Braze email campaign
+
+[6.6.7] - 2026-03-04
+---------------------
+* feat: expose admin list endpoint with search and pagination
+
+[6.6.6] - 2026-02-20
+---------------------
+* feat: Implement soft delete for admins
+
 [6.6.5] - 2026-02-24
 ---------------------
 * build: remove pinned pip now that pip-tools supports pip 26.0